Move existing code into a 'common' sub-crate
